requirements.txt
openai
langchain
langchain-experimental
streamlit
streamlit-chat
streamlit-extras
altair==4.0
python-dotenv
watchdog  # to supress warning
pre-commit
st-pages
plotly
tabulate
